# Inline Generator class