--- a/src/main/php/util/data/Generator.class.php
+++ b/src/main/php/util/data/Generator.class.php
@@ -4,6 +4,7 @@
  * A generator produces an infinite sequence of data, like for example
  * `/dev/urandom` on Unix-like operating systems.
  *
+ * @deprecated
  * @test  xp://util.data.unittest.GeneratorTest
  */
 class Generator extends \lang\Object implements \Iterator {