--- a/src/test/php/util/data/unittest/GeneratorTest.class.php
+++ b/src/test/php/util/data/unittest/GeneratorTest.class.php
@@ -2,6 +2,7 @@
 
 use util\data\Generator;
 
+/** @deprecated */
 class GeneratorTest extends \unittest\TestCase {
 
   /**