(*Like #39 but for `generate()` and `iterate()` creation methods)*

Measurable:

```php
use util\data\Sequence;

class Operations extends \util\profiling\Measurable {

  #[@measure]
  public function generate() {
    return Sequence::generate('rand')
      ->limit(10000)
      ->each()
    ;
  }

  #[@measure]
  public function iterate() {
    return Sequence::iterate(0, function($i) { return $i+= 2; })
      ->limit(10000)
      ->each()
    ;
  }
}
```

Results:

```bash
# Before
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
generate: 1000 iteration(s), 5.367 seconds, result= 10000
iterate: 1000 iteration(s), 6.571 seconds, result= 10000

# After
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
generate: 1000 iteration(s), 1.908 seconds, result= 10000
iterate: 1000 iteration(s), 2.243 seconds, result= 10000
```

The speedup is between 2.8 and 2.9 for both methods